refactor: move most of the viewer code to separate hook
